# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Taller Negreira API
  description: API para gestión de taller mecánico (Clientes, Vehículos, Órdenes, Inventario y Pagos).
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Servidor de Desarrollo Local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PaymentRequest:
      type: object
      required:
        - orden_id
        - monto
        - metodo_pago
      properties:
        orden_id:
          type: integer
        monto:
          type: number
          example: 50.00
        metodo_pago:
          type: string
          enum: [Efectivo, QR, Tarjeta]
        referencia:
          type: string
          example: REF-123456
paths:
...
  /payments:
    post:
      summary: Registrar Pago
      tags: [Payments]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentRequest"
      responses:
        "201":
          description: Pago registrado
  /payments/history:
    get:
      summary: Historial de Pagos
      tags: [Payments]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista histórica de transacciones

  # --- SERVICES ---
...